Рефакторинг EditorManager, фикс метки начального состояния (#122)
