+++ src/renderer/src/lib/data/EditorManager/index.ts
@@ -1,0 +1,1 @@
+export { EditorManager } from './EditorManager';
